move parser utils to parser
