+++ aito/cli/sub_commands/infer_table_schema_sub_command.py
@@ -6,8 +6,7 @@ from typing import List
 from aito.schema import AitoTableSchema
 from aito.utils.data_frame_handler import DataFrameHandler
 from .sub_command import SubCommand
-from ..parser import InputArgType, ParseError
-from ..parser_utils import create_sql_connecting_from_parsed_args
+from ..parser import InputArgType, ParseError, create_sql_connecting_from_parsed_args
 
 
 class InferFromFormatSubCommand(SubCommand):
